requirements.txt
absl-py==0.10.0
aio-pika==6.8.0
aiofiles==0.6.0
aiohttp==3.6.3
aiormq==3.3.1
appdirs==1.4.4
APScheduler==3.7.0
apturl==0.5.2
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
azure-core==1.12.0
azure-eventhub==5.2.1
azure-iot-device==2.5.0
azure-iot-hub==2.2.3
beautifulsoup4==4.6.0
bidict==0.21.2
blinker==1.4
boto3==1.17.30
botocore==1.20.30
Brlapi==0.6.6
cachetools==4.2.1
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorclass==2.2.0
coloredlogs==14.3
colorhash==1.0.3
contextvars==2.4
cryptography==3.4.6
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.22
dataclasses==0.8
decorator==4.4.2
defer==1.0.6
deprecation==2.1.0
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
dlib==19.21.0
dm-tree==0.1.5
dnspython==1.16.0
docopt==0.6.2
fbmessenger==6.0.0
feedparser==5.2.1
filelock==3.0.12
future==0.18.2
futures==3.1.1
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
graphsurgeon==0.4.5
grpcio==1.36.1
h11==0.9.0
h5py==2.10.0
html5lib==0.999999999
httpcore==0.11.1
httplib2==0.9.2
httptools==0.1.1
httpx==0.15.4
humanfriendly==9.1
idna==2.6
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==3.7.3
importlib-resources==5.1.2
isodate==0.6.0
janus==0.4.0
Jetson.GPIO==2.0.16
jmespath==0.10.0
joblib==0.15.1
jsonpickle==1.5.2
jsonschema==3.2.0
kafka-python==2.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==1.0
matplotlib==3.3.4
mattermostwrapper==2.2
msrest==0.6.21
multidict==4.7.6
networkx==2.5
nltk==3.5
numpy==1.16.1
oauth==1.0.1
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.45.1
onboard==1.4.1
opt-einsum==3.3.0
packaging==20.9
paho-mqtt==1.5.1
pamqp==2.3.0
pandas==1.2.3
pathlib==1.0.1
pbr==5.5.1
Pillow==8.1.2
prompt-toolkit==2.0.10
protobuf==3.15.6
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydot==1.4.2
pygobject==3.26.1
PyJWT==2.0.1
pykwalify==1.8.0
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
pyrsistent==0.17.3
PySimpleGUI==4.32.1
PySocks==1.7.1
pyTelegramBotAPI==3.7.6
python-apt==1.6.5+ubuntu0.5
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-debian==0.1.32
python-engineio==3.13.2
python-socketio==5.0.0
pytz==2020.5
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
questionary==1.5.2
rasa==2.3.4
rasa-sdk==2.4.0
redis==3.5.3
regex==2020.9.27
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
rfc3986==1.4.0
rocketchat-API==1.9.1
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
sanic==20.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.5.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.21.3
scipy==1.5.4
SecretStorage==2.3.1
sentry-sdk==0.19.5
seqeval==1.2.2
simplejson==3.13.2
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
sniffio==1.2.0
SQLAlchemy==1.3.23
ssh-import-id==5.7
stevedore==3.3.0
system-service==0.3
systemd-python==234
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.1+nv20.12
tensorflow-addons @ file:///home/thesis/Documents/jetson_setup/addons/artifacts/tensorflow_addons-0.12.0-cp36-cp36m-linux_aarch64.whl
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-probability==0.11.1
tensorflow-text @ file:///home/thesis/Documents/jetson_setup/text/tensorflow_text-2.3.0-cp36-cp36m-linux_aarch64.whl
tensorrt==7.1.3.0
termcolor==1.1.0
terminaltables==3.1.0
testresources==2.0.1
threadpoolctl==2.1.0
tk==0.1.0
tqdm==4.56.2
twilio==6.50.1
typeguard==2.11.1
typing-extensions==3.7.4.3
tzlocal==2.1
uamqp==1.2.14
ubuntu-drivers-common==0.0.0
uff==0.6.9
ujson==4.0.2
underthesea==1.2.3
Unidecode==1.2.0
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.25.11
urwid==2.0.1
uvloop==0.14.0
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.2
wcwidth==0.2.5
webencodings==0.5
webexteamssdk==1.6
websockets==8.1
Werkzeug==2.0.0rc2
wrapt==1.12.1
xkit==0.0.0
yarl==1.5.1
youtube-dl==2018.3.14
zipp==3.4.1
zope.interface==4.3.2
